Paramitise ilm template
Currently the template has hardcoded values which has to be used by all environments.
We would expect Live to differ to development and staging as the retention of logs is a requirement.

By paramitising the template we can have different values set for all environments making it more dynamic and easier to manage.

Resolves: DVOP-1832