Split root and data diskspace graphs
Split the graphs as a pre-requisite to adding averages and alerting.

Changes also include naming conventions and adding scale maximums to memory and disk graphs so can see how much capacity it has in full. These are to be configured in the pipeline as the values changes per environment.

Partially resolves: DVOP-1962