Configure slack notification via API
Configuration and API request to setup the slack notification channel which will be used to alert in Grafana.

Checks to see if slack_notification exists and if it doesn't it will create. If already exists then will skip the creation task. Post command will fail if already exists which is the reasons for the check.

Slack webhook needs to be provided by the pipeline and will be stored in vault.

Partially resolves: DVOP-1965