Add alerting to grafana
Adds alerting to grafana for cpu, root disk and data disk.
CPU will alert to slack when above the configured threshold and the disk space for root and data will alert when below the configured threshold.

The threholds are set via the pipeline as they need to differentiate between the different logging stack environments.

Partially resolves: DVOP-1964