Add cluster lb for logstash connectivity
Add a new loadbalancer that will be used by the logstash templates. This will fix the issue when we rebuild logging stack going forward as the current setup looksup ip of the ingest nodes and requires logstash playbook to be run to pick up the new instances. The lb is internal and is only accessible by logstash on http and port 9200.

Renames the existing loadbalancer for consistency and targets the master nodes instead of the cold.